move: 클래스 위치 변경에 다른 변경
