+++ src/main/java/com/twentyone/steachserver/domain/classroom/dto/UpComingClassRooms.java
@@ -1,5 +1,6 @@
-package com.twentyone.steachserver.domain.classroom.model;
+package com.twentyone.steachserver.domain.classroom.dto;
 
+import com.twentyone.steachserver.domain.classroom.model.Classroom;
 import lombok.AccessLevel;
 import lombok.Setter;
