+++ src/main/java/com/twentyone/steachserver/domain/member/service/StudentService.java
@@ -1,6 +1,9 @@
 package com.twentyone.steachserver.domain.member.service;
 
+import com.twentyone.steachserver.domain.member.model.Student;
+
 import java.util.Optional;
 
 public interface StudentService {
+    public Optional<Student> findStudentById(Integer id);
 }
