+++ src/main/java/com/twentyone/steachserver/domain/member/service/StudentServiceImpl.java
@@ -1,5 +1,7 @@
 package com.twentyone.steachserver.domain.member.service;
 
+import com.twentyone.steachserver.domain.member.model.Student;
+import com.twentyone.steachserver.domain.member.repository.StudentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -8,6 +10,7 @@ import java.util.Optional;
 @Service
 @RequiredArgsConstructor
 public class StudentServiceImpl implements StudentService{
+    private final StudentRepository studentRepository;
     @Override
     public Optional<Student> findStudentById(Integer id) {
         return studentRepository.findById(id);
